// src/types/employee.ts
export interface Employee {
  id: string
  firstName: string
  lastName: string
  email?: string
  extension?: string
  extensionNumber?: string
  phoneNumber?: string
  did?: string
  location: string
  team: string
  title?: string
  jobTitle?: string
  department?: string
  photoUrl?: string
  avatarUrl?: string
}